Conflicts:
	src/server/game/Scripting/ScriptLoader.cpp
	src/server/scripts/Northrend/Nexus/Oculus/boss_varos.cpp
	src/server/scripts/Northrend/Nexus/Oculus/instance_oculus.cpp
	src/server/scripts/Northrend/Nexus/Oculus/oculus.h

+++ src/server/game/Events/GameEventMgr.cpp
@@ -298,7 +298,7 @@ void GameEventMgr::LoadFromDB()
             if (mGameEvent[event_id].state != GAMEEVENT_NORMAL && mGameEvent[event_id].state != GAMEEVENT_INTERNAL)
             {
                 mGameEvent[event_id].state = (GameEventState)(fields[1].GetUInt8());
-                mGameEvent[event_id].nextstart    = time_t(fields[2].GetUInt64());
+                mGameEvent[event_id].nextstart    = time_t(fields[2].GetUInt32());
             }
             else
             {
